fixed updates; added clear cache
